Fixes protobuf related warnings (#995)
* 1. build swift protobuf for distribution
2. ignore warnings in protobuf lib
3. enable -Wshorten-64-to-32
4. fixes some unused variable warnings

+++ src/Ethereum/RLP.cpp
@@ -222,7 +222,7 @@ RLP::DecodedItem RLP::decode(const Data& input) {
     if (prefix <= 0xbf) {
         // long string
         auto lenOfStrLen = prefix - 0xb7;
-        auto strLen = decodeLength(subData(input, 1, lenOfStrLen));
+        auto strLen = static_cast<size_t>(decodeLength(subData(input, 1, lenOfStrLen)));
         if (inputLen < lenOfStrLen || inputLen < lenOfStrLen + strLen) {
             throw std::invalid_argument("Invalid rlp encoding length");
         }
@@ -254,7 +254,7 @@ RLP::DecodedItem RLP::decode(const Data& input) {
     } 
     if (prefix <= 0xff) {
         auto lenOfListLen = prefix - 0xf7;
-        auto listLen = decodeLength(subData(input, 1, lenOfListLen));
+        auto listLen = static_cast<size_t>(decodeLength(subData(input, 1, lenOfListLen)));
         if (inputLen < lenOfListLen || inputLen < lenOfListLen + listLen) {
             throw std::invalid_argument("Invalid rlp list length");
         }
