+++ tests/interface/TWTestUtilities.h
@@ -40,21 +40,21 @@ std::string getTestTempDir(void);
             auto inputData = input.SerializeAsString();\
             auto inputTWData = WRAPD(TWDataCreateWithBytes((const uint8_t *)inputData.data(), inputData.size()));\
             auto outputTWData = WRAPD(TWAnySignerSign(inputTWData.get(), coin));\
-            output.ParseFromArray(TWDataBytes(outputTWData.get()), TWDataSize(outputTWData.get()));\
+            output.ParseFromArray(TWDataBytes(outputTWData.get()), static_cast<int>(TWDataSize(outputTWData.get())));\
         }
 #define ANY_ENCODE(input, coin) \
         {\
             auto inputData = input.SerializeAsString();\
             auto inputTWData = WRAPD(TWDataCreateWithBytes((const uint8_t *)inputData.data(), inputData.size()));\
             auto encodedData = WRAPD(TWAnySignerEncode(inputTWData.get(), coin));\
-            encoded = TW::data(TWDataBytes(encodedData.get()), TWDataSize(encodedData.get()));\
+            encoded = TW::data(TWDataBytes(encodedData.get()), static_cast<int>(TWDataSize(encodedData.get())));\
         }
 #define ANY_PLAN(input, output, coin) \
         {\
             auto inputData = input.SerializeAsString();\
             auto inputTWData = WRAPD(TWDataCreateWithBytes((const uint8_t *)inputData.data(), inputData.size()));\
             auto outputTWData = WRAPD(TWAnySignerPlan(inputTWData.get(), coin));\
-            output.ParseFromArray(TWDataBytes(outputTWData.get()), TWDataSize(outputTWData.get()));\
+            output.ParseFromArray(TWDataBytes(outputTWData.get()), static_cast<int>(TWDataSize(outputTWData.get())));\
         }
 #define DUMP_PROTO(input) \
         { \
